refactor: rearrange type def
