+++ packages/readr/pages/about.tsx
@@ -1,5 +1,6 @@
 // under construction
 
+import { ApolloError } from '@apollo/client/core'
 import errors from '@twreporter/errors'
 import type { GetServerSideProps } from 'next'
 import { ReactElement, useMemo, useState } from 'react'
@@ -9,8 +10,8 @@ import client from '~/apollo-client'
 import Awards from '~/components/about/awards'
 import Landing from '~/components/about/landing'
 import LayoutWithLogoOnly from '~/components/layout/layout-with-logo-only'
-import { awards as awardsGql } from '~/graphql/query/award'
-import type { Award, Language, RenderedAward } from '~/types/about'
+import { Award, awards as awardsGql } from '~/graphql/query/award'
+import type { Language, RenderedAward } from '~/types/about'
 
 import type { NextPageWithLayout } from './_app'
 
@@ -24,6 +25,7 @@ type languageWording = {
     title: string
     content: string
   }
+  awardsTitle: string
 }
 
 const wording: Record<Language, languageWording> = {
@@ -33,6 +35,7 @@ const wording: Record<Language, languageWording> = {
       content:
         'READr 是致力於以<b>資料</b>做<b>新聞</b>與<b>內容實驗</b>的媒體。於 2018 年正式創立，重視與各種專業者、讀者的協作，期望讓以往封閉的新聞編輯室有開放的可能。<b>資料分析、多媒體互動、理想的閱讀體驗、開放資料、開源工具</b>，都是我們提供的服務。',
     },
+    awardsTitle: '獲獎經歷',
   },
   en: {
     landing: {
@@ -40,6 +43,7 @@ const wording: Record<Language, languageWording> = {
       content:
         'READr uses <b>data</b> for <b>news</b> and <b>content</b> <b>experimentation</b>. Collaboration with diverse professionals and readers opens the once-closed newsroom. We provide services like <b>data</b> <b>analysis,</b> <b>multimedia,</b> <b>ideal</b> <b>reading</b> <b>experiences,</b> <b>open</b> <b>data,</b> <b>and</b> <b>tools.</b>',
     },
+    awardsTitle: 'Awards',
   },
 }
 
@@ -107,16 +111,24 @@ const About: NextPageWithLayout<PageProps> = ({ awardsData }) => {
       />
       <Awards
         renderedAwards={renderedAwards}
-        title={language === 'ch' ? '獲獎經歷' : 'Awards'}
+        title={wording[language].awardsTitle}
       />
     </Page>
   )
 }
 
+interface QueryResult {
+  data: {
+    awards: Award[]
+  }
+  loading: boolean
+  error?: ApolloError
+}
+
 export const getServerSideProps: GetServerSideProps<PageProps> = async ({}) => {
   let awardsData: Award[] = []
   try {
-    const result = await client.query({
+    const result: QueryResult = await client.query({
       query: awardsGql,
     })
     awardsData = result.data.awards
